fuzzer: re-sync trinity files
Move them to their own directory so it's more obvious where they come
from

+++ fuzzer/trinity_files/include/sanitise.h
@@ -3,7 +3,7 @@
 
 #include "syscall.h"
 
-unsigned long long rand64(void);
+u64 rand64(void);
 
 void sanitise_mmap(int childno);
 void sanitise_rt_sigaction(int childno);
